requirements.txt
sentry-sdk==2.27.0
google-api-python-client==2.168.0
gspread==6.2.0
django_formtools==2.5.1
django-jsonform==2.23.2